Simplify DEFAULT_SESSION_LIFETIME type casting

+++ src/service/session/mod.rs
@@ -119,7 +119,7 @@ impl Timer {
 }
 
 /// Default session lifetime in seconds (10 minutes)
-const DEFAULT_SESSION_LIFETIME: u64 = 600;
+const DEFAULT_SESSION_LIFETIME: u32 = 600;
 
 /// turn session information.
 ///
@@ -370,7 +370,7 @@ where
                     // A random string of length 16.
                     nonce: generate_nonce(),
                     // Current time stacks for DEFAULT_SESSION_LIFETIME seconds.
-                    expires: self.timer.get() + DEFAULT_SESSION_LIFETIME,
+                    expires: self.timer.get() + DEFAULT_SESSION_LIFETIME as u64,
                 },
             );
         }
@@ -476,7 +476,7 @@ where
                 Session::Authenticated {
                     allocate_channels: Vec::with_capacity(10),
                     permissions: Vec::with_capacity(10),
-                    expires: self.timer.get() + DEFAULT_SESSION_LIFETIME,
+                    expires: self.timer.get() + DEFAULT_SESSION_LIFETIME as u64,
                     username: username.to_string(),
                     allocate_port: None,
                     password,
@@ -578,7 +578,7 @@ where
 
             // Records the port assigned to the current session and resets the alive time.
             let port = self.port_allocator.lock().alloc(None)?;
-            *expires = self.timer.get() + (lifetime.unwrap_or(DEFAULT_SESSION_LIFETIME as u32) as u64);
+            *expires = self.timer.get() + lifetime.unwrap_or(DEFAULT_SESSION_LIFETIME) as u64;
             *allocate_port = Some(port);
 
             // Write the allocation port binding table.
